add : capability to test server-node store with testStore utility
